Use grid to position the puzzle elements
Changed from table to grid layout. The initial file to display is now the conventional index.html . Added a docker-compose for launching an instance of nginx. Also tried to make the display a little more mobile friendly by having the explanatory text fold under the puzzle on small "devices" (page widths). A future improvement may be to use vmin and vmax to scale the puzzle.